+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Unreleased` header.
 # Unreleased
 
 - On macOS, remove spurious error logging when handling `Fn`.
+- On X11, fix an issue where floating point data from the server is
+  misinterpreted during a drag and drop operation.
 
 # 0.29.4